BackEnd/requirements.txt
Django==5.2.7
django-mongodb-backend==0.1.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
pymongo==4.10.1

